[MODIFY] <FirstView> 루트 변경 & 예외 처리
